requirements.txt
pydantic>=2.0.0
google-genai
python-dotenv
psycopg2-binary
